+++ bm/Search/SearchViewController.swift
@@ -38,7 +38,7 @@ class SearchViewController: UIViewController, UISearchBarDelegate {
         if #available(iOS 13.0, *) {
             searchBar?.searchTextField.backgroundColor = .white
             searchBar?.searchTextField.textColor = .black
-            searchBar?.barStyle = .black
+            searchBar?.overrideUserInterfaceStyle = .light
         }
 
         searchBar?.backgroundColor = .white